Remove Falcon from fp16 example for now

+++ python/llm/example/GPU/HF-Transformers-AutoModels/Model/falcon/generate.py
@@ -47,7 +47,7 @@ if __name__ == '__main__':
                                                  load_in_4bit=True,
                                                  trust_remote_code=True,
                                                  use_cache=True)
-    model = model.half().to('xpu')
+    model = model.to('xpu')
 
     # Load tokenizer
     tokenizer = AutoTokenizer.from_pretrained(model_path,
